feat: enhance BookSerie show page with related books data and update types in migrations

+++ database/migrations/2026_08_02_041321_create_games_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
             $table->string('background_public_id')->nullable();
             $table->string('developer')->nullable();
             $table->string('publisher')->nullable();
-            $table->integer('rating')->nullable();
+            $table->unsignedTinyInteger('rating')->nullable();
             $table->boolean('hg')->default(false);
             $table->string('version')->nullable();
             $table->timestamps();
